updated naive to long

+++ challenges/data-structures/arrays/algorithmic-crush/java8/failed-naive/Solution.java
@@ -8,20 +8,20 @@ import java.util.*;
 
 public class Solution {
 
-	/*
-	Problem:
-	You get two integers, N and M. Make a list of zeroes of size N.
-	There are M commands that follow, of the format:
-	a b k
-	Meaning, add the number k to all integers at indices in the range a-b inclusive. (1-indexed!)
-	Once you're done, spit out the biggest integer in the N-size list.
-	*/
+    /*
+    Problem:
+    You get two integers, N and M. Make a list of zeroes of size N.
+    There are M commands that follow, of the format:
+    a b k
+    Meaning, add the number k to all integers at indices in the range a-b inclusive. (1-indexed!)
+    Once you're done, spit out the biggest integer in the N-size list.
+    */
 
-	// Read input scanner from system.in
-	private static Scanner in = new Scanner(System.in);
-	private static int[] list;
+    // Read input scanner from system.in
+    private static Scanner in = new Scanner(System.in);
+    private static long[] list;
 
-	public static void main(String[] args) {
+    public static void main(String[] args) {
         // Initialize the list
         int size = in.nextInt();
         initList(size);
@@ -29,8 +29,8 @@ public class Solution {
         // Read commands
         int commandCount = in.nextInt();
         for (int i = 0; i < commandCount; i++) {
-        	// Read in the command, swapping to zero indexed, inclusive-exclusive.
-        	add(in.nextInt() - 1, in.nextInt(), in.nextInt());
+            // Read in the command, swapping to zero indexed, inclusive-exclusive.
+            add(in.nextInt() - 1, in.nextInt(), in.nextInt());
         }
 
         System.out.println(getMax());
@@ -40,7 +40,7 @@ public class Solution {
     * @param size The size of the list to initialize
     */
     private static void initList(int size) {
-    	list = new int[size];
+        list = new long[size];
     }
 
     /** Adds a value onto a specified range of the list
@@ -49,7 +49,7 @@ public class Solution {
     * @param b The last index to be processed
     */
     private static void add(int a, int b, int val) {
-    	for (int i = a; i < b; i++) {
+        for (int i = a; i < b; i++) {
             list[i] += val;
         }
     }
@@ -57,8 +57,8 @@ public class Solution {
     /** Finds the maximum value in the list
     * @returns The maximum value in the ranged list
     */
-    private static int getMax() {
-    	int max = Integer.MIN_VALUE;
+    private static long getMax() {
+        long max = Long.MIN_VALUE;
         for (int i = 0; i < list.length; i++) {
             if (list[i] > max) {
                 max = list[i];
